+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased (v0.3.3)
-* Show more info auto-encode fails to find a suitable crf.
+* Show more info when auto-encode fails to find a suitable crf.
 
 # v0.3.2
 * Improve sample generation speed & frame duration accuracy.